edited task update_advertisement_task, signals create_date_of_deactivate_and_delete, model advertisement

+++ api_domer/tasks.py
@@ -2,6 +2,7 @@ import os
 from celery import shared_task
 from django.core.files import File
 from django.db import transaction
+from django.forms import model_to_dict
 from django.template.loader import render_to_string
 
 from advertisement.models import Field, PhotoAdvertisement, Advertisement
@@ -43,8 +44,8 @@ def save_advertisement_task(user, data, additional_information, processed_photo)
 
             if not os.listdir(folder_path):  # Если папка пуста
                 os.rmdir(folder_path)  # Удаляем папку
-    except Exception as e:
 
+    except Exception as e:
         if processed_photo:
             preview_img = processed_photo.get('preview_img')
             other_images = processed_photo.get('other_img', None)
@@ -58,17 +59,15 @@ def save_advertisement_task(user, data, additional_information, processed_photo)
             if not os.listdir(folder_path):  # Если папка пуста
                 os.rmdir(folder_path)  # Удаляем папку
 
+        # !!! Переписать на новую функцию отправки писем!!!
         html_content = render_to_string(
             "asend_create_advertisement_error.html",
             context={"activation_title": data['title'], "url": env_keys.get("URL")},
         )
-
         send_email_task(chat_title='Ошибка при создании обьявления',
                         recipient=data['email'],
                         text_content=html_content,
                         html_content=html_content)
-
-
 
 
 @shared_task()
@@ -80,50 +79,59 @@ def update_advertisement_task(user, advertisement_id, data, additional_informati
     for field in additional_information_save:
         additional_information[field.title] = ', '.join(additional_information.pop(f'{field.id}'))
 
-    Advertisement.objects.filter(author=user,
-                                 id=advertisement_id
-                                 ).update(moderated='',
-                                          additional_information=additional_information,
-                                          **data,
-                                          additional_information_view=list(additional_information.items()),
-                                          is_active=False)
+    editing_advertisement = Advertisement.objects.get(author=user, id=advertisement_id)
 
-    advertisement = Advertisement.objects.get(author=user, id=advertisement_id)
-    advertisement.save()
+    data['moderated'] = ''
+    data['additional_information'] = additional_information
+    data['is_active'] = False
+
+    for key, value in data.items():
+        setattr(editing_advertisement, key, value)
+
+    if not new_preview_photo_from_old_ones and not processed_photo and not delete_photo:
+        editing_advertisement.save()
+        return
 
     if new_preview_photo_from_old_ones:
-        old_preview_image = advertisement.preview_image
+        print('Меняем две фотки местами')
+        old_preview_image = editing_advertisement.preview_image
         old_photo = PhotoAdvertisement.objects.get(photo=new_preview_photo_from_old_ones,
-                                                   advertisement=advertisement).photo
+                                                   advertisement=editing_advertisement)
 
-        Advertisement.objects.filter(author=user, id=advertisement_id).update(preview_image=old_photo)
-        PhotoAdvertisement.objects.filter(photo=new_preview_photo_from_old_ones,
-                                          advertisement=advertisement).update(photo=old_preview_image)
+        editing_advertisement.preview_image = old_photo.photo
+        editing_advertisement.save()
+
+        old_photo.photo = old_preview_image
+        old_photo.save()
+
     if processed_photo:
-        preview_img = processed_photo.get('preview_img', None)
+        print('Добавляем новые фотки и можем поменять местами')
+        preview_img = processed_photo.get('preview_img')
         other_images = processed_photo.get('other_img', None)
         if preview_img:
-            if advertisement.preview_image:
-                PhotoAdvertisement.objects.create(photo=advertisement.preview_image, advertisement=advertisement)
+            if editing_advertisement.preview_image:
+                PhotoAdvertisement.objects.create(photo=editing_advertisement.preview_image,
+                                                  advertisement=editing_advertisement)
             with open(preview_img, 'rb') as f:
-                advertisement.preview_image = File(f)
-                advertisement.save()
+                editing_advertisement.preview_image = File(f)
+                editing_advertisement.save()
             os.remove(preview_img)
         if other_images:
             for photo in other_images:
                 with open(photo, 'rb') as f:
-                    additional_photo = PhotoAdvertisement(photo=File(f), advertisement=advertisement)
+                    additional_photo = PhotoAdvertisement(photo=File(f), advertisement=editing_advertisement)
                     additional_photo.save()
                 os.remove(photo)
 
     if delete_photo:
-        PhotoAdvertisement.objects.filter(photo__in=delete_photo, advertisement=advertisement).delete()
-        if advertisement.preview_image in delete_photo:
-            file_path = advertisement.preview_image.path
+        print('Удаляем фотки')
+        PhotoAdvertisement.objects.filter(photo__in=delete_photo, advertisement=editing_advertisement).delete()
+        if editing_advertisement.preview_image in delete_photo:
+            file_path = editing_advertisement.preview_image.path
             folder_path = os.path.dirname(file_path)
             os.remove(file_path)
-            advertisement.preview_image = None
-            advertisement.save()
+            editing_advertisement.preview_image = None
+            editing_advertisement.save()
 
             if not os.listdir(folder_path):  # Если папка пуста
                 os.rmdir(folder_path)  # Удаляем папку
